requirements.txt
apache-beam==2.23.0
avro-python3==1.9.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
crcmod==1.7
dill==0.3.0
docopt==0.6.2
fastavro==0.21.24
future==0.18.2
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-pubsub==1.0.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
mock==2.0.0
mysql-connector-python==8.0.17
numpy==1.17.4
oauth2client==3.0.0
oauthlib==3.1.0
pbr==5.4.3
protobuf==3.10.0
pyarrow==0.14.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pydot==1.4.1
pymongo==3.9.0
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.13
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tweepy==3.8.0
urllib3==1.25.6